mid-test fixes

+++ clc/modules/msgs/src/main/java/edu/ucsb/eucalyptus/msgs/Messages.groovy
@@ -376,7 +376,7 @@ public class NetworkConfigType extends EucalyptusData {
   String ignoredPublicIp;
   String privateDnsName;
   String publicDnsName;
-  Long vlan;
+  Integer vlan;
   Long networkIndex;
   
   def NetworkConfigType() {
